Pass kwarg values directly to indy call site
This is a first step toward passing keyword arguments all the way
through the stack without allocating a Hash object. When we detect
that there's a trailing Hash operand in the argument list, we know
it must be a keyword arguments hash. Rather than eagerly creating
the Hash, we instead evaluate its values as normal but leave them
on the stack for the call. The invocation takes the positional and
keyword argument values along with a keyword names descriptor and
reassembles the values into a kwargs hash lazily before invoking
a normal call site.

Once the normal call process has a way to plumb kwarg values all
the way through to a target method, this will push further into
that stack.