Direct kwargs are too hard without indy
This is the first place where the Indy compiler will start to
seriously diverge from the non-Indy version. It's just too hard
to plumb direct keyword arguments through the non-Indy compiler, so
we add a way for the IR visitor to alter compilation strategy if
running in a mode that has reduced capabilities.