Make constant visibility autoload-aware
Updating the visibility of an constant being defined for autoload
caused that autoload constant to be set to UNDEF without the new
visibility, because the logic did not handle UNDEF values (our
marker for an autoloading constant).

The changes here check for undef and update the autoload instead
of the constant table. Also modified is tweaking Autoload to only
update the value if the incoming newValue is not null or UNDEF,
which may happen due to the way that constant visibility updates
happen.

Fixes one failure in Zeitwerk relating to private autoloading
constants.