Invoke const_added for Autoload update
When a constant is currently being autoloaded, we should still
invoke const_added to indicate that someone set it. This was the
cause of most Zeitwerk problems: an autoload gets registered for
the constant associated with the current file, and when that
constant gets assigned it should trigger further autoload
registrations for e.g. subdirs of a gem.