respond_to?(<symbol>) optimization
This optimizes the common case of respond_to? with a literal
symbol. When the method is "naturally" respond_to? => true then
that result can be cached behind the usual guards. This makes the
success path essentially free.

The logic here has a lot of copied behavior from elsewhere and
utilizes some now-static methods from InvokeSite, so it could be
refactored along with the rest of indy logic to be much simpler.