Handle empty keywords for Time.new
Solves https://github.com/jruby/jruby/issues/8854

Prevents `Time.new(2023, 2, 1, **{})` from raising NullPointerException
when accessing `opts` which can happened when keyword args are empty, e.g:
https://github.com/rails/rails/blob/v7.1.5.1/activesupport/lib/active_support/testing/time_helpers.rb#L185

Unsure if this is the best way, but it does the job for now I think.

I realised that if keywords was empty, `hasKeywords` still returns true
and `extractKeywordArgs` returns null, and `args` doesn't have the Hash object,
it is just missing. So that is why we can only subtract 1 from
argc after `extractKeywordArgs` and confirming it is not null.

It feels there is a more general problem of how keywords are handled or/and
a missing helper to avoid these issues as I think this bug may be in a
couple more places in the codebase, but probably people don't hit them frequently.

I added a test for it since it was a regression from JRuby v9, but also
unsure where best to put the test or if it is something we want / need.

Feel free to make any changes or suggestions.