Use jruby-base sources for jruby-complete
jruby-complete was previously bundling jruby-core sources to
fulfull the requirement of maven central for sources to be
provided, but that artifact only published a dummy source jar. In
this snapshot work, for some reason that jruby-core source jar is
no longer created, but then jruby-complete has no sources to copy.

This change modifies jruby-complete to use jruby-base sources,
which provides the sources for JRuby itself as it should.

I am unclear on why this worked in the past, but it could be
something related to the old parent pom we no longer use. What I
have here works, but I seek a better explanation for why it broke.