Make this zero-alloc and fix null check
The null check here appears to be intended in the CRuby code to
check whether the given encoding name has already been set, and
return true if the new encoding would overwrite it. This is not
currently consumed by any code.

The switch to a lambda here avoids allocating the def_p out array.
The value originally put in that array is only used to set either
the default internal or default external encoding, which can be
done in a lambda.

Part of improvements for #8840