Ensure CFG is not null or raise
Both branches of the CFG null check below required the CFG to be
non-null, the second via the call to fic.toStringInstrs. Since we
have not seen NPE here, the CFG appears to never be null in our
uses of this listener, so ensure that is true with a hard check
and remove the null path.

Part of fixes for #8840