Also ensure we are using longs for bits here
This went unfixed during the original long bitfield fixes in
4b897eec13b4 and was not noticed until I started playing with the
hexapdf benchmarks again in #8824. This only breaks when the total
number of elements (strings and dynamic objects) exceeds 31, since
the integer shift will then produce zero and stop setting higher
bits. It also does not happen for dstr with more than 50 elements,
the limit of the current optimization.