Don't search Object for const_defined? pathing
The change here is form-fit to this specific case, but basically,
we never search Object as an inherited namespace for const_defined?
and if we are doing a namespaced lookup don't do the extra search
against Object either.

Other paths like const_get should continue to search Object to fit
the spec.

Fixes #8867