Do not exclude bin/jruby and bin/ruby from artifacts
This exclude was formerly countered by an additional copy of
bin/jruby.sh to bin/jruby and bin/ruby, but when we added new
bin/jruby and bin/ruby scripts to the repository and removed that
copying, the excludes took effect and prevented those scripts from
being included in any of the artifacts dependent on the "complete"
layout. This appears to include jruby-dist, which obviously must
have those scripts to be useful.

I remove the excludes here, since these files are now no longer
copied and should just be included as-is from the repo's bin
directory. This appears to fix the missing files in the dist
archive.