Move NthRef logic to BuildBackref instr
NthRef has the side effect requirement of a backref frame, so it
should not be represented as a side-effect free operand. This
patch introduces BuildNthRefInstr due to requirements that high
group numbers warn, which conflicts with the lettered backrefs
used in BuildBackrefInstr.

The lack of a frame caused optimized NthRef logic in the JIT to
fail to push a frame, which caused those references to break once
JIT had settled.

This patch also fixes defined?(any backref form) to also trigger a
backref frame, also broken once JIT settled.