Go back to static require for java.sql
If we get a NCDFE attempting to reflect against the java.sql
DriverManager, it's pretty unlikely we were able to load anything
into it either. If we didn't load it, we don't care if we can't
clear it.

This allows us to move java.sql back to requires static since it is
only needed when used.