Test loadedFeatureWithPath with a simple feature lock name
When a simple feature lock name is provided here, such as when a
thread has just started doing its search for a given feature, the
initial backward walk through that name used to walk past the first
character. This subsequently triggered an AIOOB exception.

The test here attempts to simulate this situation, since it's a
very tricky set of concurrent loads required to trigger it
organically.

See https://github.com/jruby/jruby/issues/8958