Strip scheme at first and append as needed
This fixes one issue with the new logic to process globs without
scheme. The intent of this code appears to have been to remove the
scheme from the path portion, used for fnmatching and character-by-
character comparisons. However in that case the scheme must always
be prepended when querying the "filesystem", so that URLs and such
can always use appropriate logic. The patch here does the
following:

* Fully remove the scheme from the passed-in path, since the
  glob_helper logic seems to assume that the path bytes provided
  will be the schemeless full path.
* Prepend the scheme when checking each link in the path, so that
  appropriate exotic filesystem checks work correctly.

This appears to fix the issue in jruby/jruby#8973.