Fix jar:file when under casefolding logic
The use of "magic" logic under macOS also appears to break the
jar:file scheme. This patch fixes it, considering both jar:file:,
the path to the jar, and the "!" to be part of the scheme.

An additional fix was required for JRubyFile.JarResource to
normalize the "." path to an empty entry, because jar:file:x.jar!.
is not a valid jar:file URI for the JDK.

Includes specs.

Relates to jruby/jruby#8973.