Remove confusing shared ByteList buf
This is trying to reuse the same bytelist instance for nested
calls to glob_helper but it is very unclear how it worked as
an infrequent viewer.  In doing this work I eliminated other
object creation so I think we splurge and just let us make
more ByteLists while walking a filesystem.