private_constant should just update existing autoload
The logic here was proceeding to set the existing autoload to
UNDEF, which then triggered it to be removed as a failed autoload.
This left us with a constant table containing a private entry of
UNDEF but no autoload in the table to load for it.

The modified logic checks if the incoming value was also UNDEF,
indicating that the value was a previously-unloaded or in-progress
autoload, and leaves the autoload table alone while just updating
the constant table with the new visibility.

Fixes jruby/jruby#8923