Add module readability to JI class binding test
We should not attempt to read public classes from modules that are
not readable from the JRuby module.

Alternative fix for #8987. Original fix in #8989 approached this
differently by requesting read access to the module. Further
discussion in the related OpenJDK core-libs-dev thread lead to the
realization that the generated proxy class from #8987 was in an
internal JDK module and not intended to be readable by anyone, but
we were attempting to find it. The fix here avoids binding classes
from modules that are not naturally readable from the JRuby module,
rather than trying to add more reads than were originally
configured or requested.