Don't attempt to add reads while binding methods
Modules that are not naturally readable should not be added to
reads just so we can get access to their methods. Any logic that
is passing unreadable modules' classes and methods to this logic
should reexamine how those classes and methods are being acquired,
and add reads at that point or avoid acquiring references to such
unreadable modules.

See #8989 for the original patch, as a fix for #8987. The OpenJDK
core-libs-dev discussion linked from that issue led us to a new fix
that avoids binding unreadable modules' classes in our Java
integration layer.