Retire safe double processing utilities
This came about when an old version of Java had a CVE and would
end up in an endless loop processing a number (numbers?).  This was
fixed in Java 1.6.

Here is a reference to it being fixed:
https://www.exploringbinary.com/fpupdater-fixes-the-java-2-2250738585072012e-308-bug/

This should get rid of multiple scans and simplify the overall codepath
of converting a string to a double/bigdecimal.