Move all main classes to org.jruby.main package
This copies org.jruby.Main and org.jruby.JarBootstrapMain to the
org.jruby.main package where we keep other "mains". This is in part
to localize mains but also to start removing classes from org.jruby
package that don't need to or should not be accessed by users. In
this case, we don't recommend using Main to execute JRuby from
within an already-running JVM, preferring our embedding APIs or
using org.jruby.Ruby directly. JarBootstrapMain is rarely (if ever)
used, so we'll also move it alongside other mains.

JarBootstrap main is marked for removal, since nobody is likely to
be using it as an API. org.jruby.Main is not marked for removal,
since it almost certainly is being used as an API and the removal
attribute might prevent that code from compiling at all.