Separate boot items to be run immediately before user code
This change primarily moves the loading of command-line requires
further into the "main" execution path, rather then running them
on instantiation of Ruby before other state has been set up.

When the command-line requires depend on other state set up by the
main path, it may behave differently to require at the command line
versus requiring in the first executed file. One example is that
no thread context classloader will be set when these files run.

Such a case was found working on a Spring Boot application. When
the Spring classes were loaded and registered before the main
script started, they would have a misconfigured SLF4J logger and
fail to boot. Moving that logic into the main script fixed the
issue, presumably because the early stages of Spring Boot's boot
process depends on the context classloader.

This moves the automatic load of command-line required files to
the main path in the method `runFromMain`. A new `boot` method is
added to include these pre-execution startup items.

A new constructor method `newMain` is added for the main path to
create a new runtime without immediately booting it, allowing for
that to happen during main execution. The other forms of
`newInstance` are modified to pre-boot the runtime, assuming it
will immediately be used to execute code.