Fix embedding that depended on global runtime
The default behavior to create a new runtime would check if a
global runtime had been intialized, and if not it called the
newInstance method expecting that the result would be registered as
the global runtime. With the changes in this PR we no longer do
this automatic registration, so the code here is changed to use
getGlobalRuntime methods that explicitly create and register the
runtime.

This exposes a risk in changing the global registration lifecycle
for the newInstance methods: any code that expects the simple act
of calling newInstance to register a global runtime will need to
be updated to avoid that expectation, either capturing the runtime
returned or calling getGlobalRuntime or useAsGlobalRuntime to
ensure that the new runtime instance is registered. I don't expect
there to be a lot of code calling these APIs directly.