Migrate .java-version contents to .jruby.release
I think it makes more sense to use .jruby.release as the only file
for version-specific details. This change may need to be tweaked
to support JRuby 10 releases prior to 10.0.3.0 (assuming it is
merged into that release), or it may not be important since we
won't be changing the jruby.sh in any released artifacts.