Refactor and cache negative results when appropriate
We can cache a true result only if respond_to? is the built-in
version and it would return true. We can cache a false result if
respond_to? is built-in and returns false and respond_to_missing?
is not define or both are undefined. All other cases must dispatch
through slow-path logic (normal dynamic invocation of respond_to?).