Make RubyArray abstract
This is a first stab at making RubyArray an abstract class, which
will allow us to represent many different storage strategies as
Array as seen from Ruby code:

* The standard implementation, based on the CRuby logic, uses a
  native Java IRubyObject[] and supports full mutability and CoW
  sharing of that array.
* Two specialized implementations extend from this and provide one-
  and two-element versions of Array without an IRubyObject[] field.
  Making RubyArray itself abstract will make it easier to add more
  specializations.
* There are many useful collections provided by the JDK that could
  be used to implement Array. This change will allow them all to
  appear as the same Ruby type but with custom behaviors for
  specific use cases.
* This will make for an easier path to thread-safe Array, since we
  can have specialized implementations that all look like Array.

This is very much a prototype. I have tried to limit the changes
throughout the codebase by making RubyArray abstract rather than
introducing a new supertype (which would have to propagate through
all RubyArray signatures in the system). The goal is to keep all
RubyArray-consuming code the same as it is today while providing
multiple backend implementations.