Remove test of automatic jar extension loading
We removed the ability to automatically load extensions from jars
many years ago, but this test remained. It only continued to pass
because some other library loaded the 'json' library which meant
that the tested classes were available. When this same example is
run outside of the test suite, it fails:

```
$ cx jruby-10.0.2.0 ruby -e "require 'uri:file:./lib/ruby/stdlib/json/ext/parser'; p JSON::Ext::Parser"
NameError: uninitialized constant JSON::Ext::Parser
  const_missing at org/jruby/RubyModule.java:4866
         <main> at -e:1
```

Because this functionality no longer exists, I am removing the
test.