Ensure we always use the smallest box for fixnums
The fixnum specialization triggered a bug in hash calculation
where something like [257,1] could hash differently depending on
when during execution it was calculated. The specific case was in
resolv.rb, which sets up a hash of integer pairs pointing at the
Class object associated with that DNS code. Later accesses hashed
the given pair differently and ended up replacing the earlier
entry.

The source of this problem is not entirely clear to me, but it was
fixed by ensuring that we never construct a cacheable fixnum value
and always use the cached values. I suspect this is a flaw in the
design of the safeRecurse logic, which uses an IdentityHashMap
when invoking `hash` on nested elements, but I cannot explain how
having two different fixnum instances could trigger this problem.

There are other issues with safeRecurse that, when fixed, would
also work around this problem (see jruby/jruby#9086) but it is at
least clear that the fixnum specialization must be "absolute"; in
other words, we must always use the cached instances for the
supported range (to ensure they retain identity) and always use the
same fixnum type for a given value range (so we don't end up with
two fixnums of the same value but different class).

These problems do not affect CRuby because fixnums are idempotent
due to their use of tagged pointers; there can never be two fixnum
objects because they are never objects.