Combine byte and short fixnums into CompactFixnum
Due to our use of an int for object flags (jruby/jruby#9088) we
have at least 16 bits that never get used in every object. This
tweak to fixnum specialization leverages those 16 bits for values
in the byte and short ranges.

This has several advantages:

* byte and short-ranged fixnums become even smaller, having no
  additional fields beyond those in RubyBasicObject.
* Fewer types means simpler construction logic and less complexity
  for the JIT.

This change might need to be reversed if at some point we choose
to shrink the size of the flags field, but for now we make good use
of it to shrink some of our numeric boxes.