Shrink varTableStamp to a byte
This stamp is used to guard atomic access to the varTable field
when it needs to be grown to accommodate more values. The logic
expects that anyone performing such an update will increment the
stamp while updating the the table reference, backing off if the
stamp has changed before finally assigning the new table (allowing
concurrent writers to all attempt growth without a hard lock, with
the first-past-the-post thread winning).

We probably don't need 32 bit for this, since that would
accommodate an absurd 2-4 billion concurrent mutators. Byte range
may be too small, but still handles at least 64 concurrent mutators
with non successfully growing the table large enough for any of the
others to proceed.

This relates to object header size reductions detailed in recent
issues like jruby/jruby#9090.