Revert frozen boolean to a bit in flags
Frozen as a boolean would still consume a byte, so even if we
eliminated the other two flags (nil and false) it would not save
us any space compared to just leaving all three in a byte-sized
flags field. And obviously without eliminating the other two flags,
we still need to store them. We may revisit this if we can show
that polymorphic "nil" and "false" tests show some benefit (and no
detriment) but ultimately we won't get smaller than 8 bits of space
for the unavoidable frozen flag.