Cache both hash forms for FString
Previously the long form of the FString was cached and returned
by longHashCode, but that value is actually just an int and can
also be returned by hashCode. We can also pre-allocate and cache
the fixnum version of the hash so it does not have to be created
each time it is used from Ruby.

Found while working on performance optimizations for Puma in
puma/puma#3838.