Never trust external content for FString
When creating an FString for the first time, we should never trust
that the incoming byte[] or ByteList are now ours to own. The
caller might still have a reference to them that they continue to
modify, resulting in a zombie FString that can't be found or that
has incorrect contents.

This came up while trying to implement the "fake string"
optimization for uncached headers in puma/puma#3838. Holding a
reference to a RubyString and its ByteList that could be updated
in-place and then used to caches new headers led to those cached
FStrings being modified directly.