*nil no longer calls nil.to_a, similar to how **nil does not call nil.to_hash
*nil no longer calls nil.to_a, similar to how **nil does not call nil.to_hash. [Feature #21047]

This could be optimized more but this matches semantics.  The missing optimization is the parser(s) need to see literal `nil` and then set a new
flag on build_splat marking that fact.  Then the helper can use a cached
empty frozen array.  The bulk of the improvement is not dispatching so this
will be a smaller percentage of improvement on a pretty uncommon case.